minivault-api/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
distlib==0.3.9
fastapi==0.104.1
filelock==3.18.0
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
packaging==25.0
pipenv==2025.0.3
platformdirs==4.3.8
pluggy==1.6.0
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
PyYAML==6.0.2
pyzipper==0.3.6
regex==2024.11.6
requests==2.31.0
safetensors==0.5.3
sniffio==1.3.1
starlette==0.27.0
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.24.0
virtualenv==20.31.2
python-dotenv==1.0.0
accelerate==0.27.2
httpx==0.27.0
